Create all necessary directories on path to build folder
If you choose to make the build directory a subdirectory of the source
folder it would previously fail to make the build directory because the
source directory did not yet exist. Now it uses os.makedirs to make the
build directory which will also create the source directory if necessary.